+++ app/src/main/java/com/example/togedy_android/data/mapper/todata/DateRequestMapper.kt
@@ -1,4 +1,4 @@
-package com.example.togedy_android.data.mapper.todata.planner
+package com.example.togedy_android.data.mapper.todata
 
 import com.example.togedy_android.data.remote.model.planner.DateRequestDto
 import com.example.togedy_android.domain.model.DateModel
